Fix how added scrollbars are detected and measured. (...)
`offsetWidth` provides the distance between the outer left border edge to the
outer right border edge. `clientWidth` provides the distance between the outer
left padding edge and the outer right padding edge - the same as `offsetWidth`,
but excluding borders and scrollbars.[^1] The difference between `offsetWidth`
and `clientWidth` includes the size of the scrollbars contributing to (or
consuming) the width.

To measure if a scrollbar was added between two different layout states, we
measure the difference between (`offsetWidth` - `clientWidth`) in each of the
two states. If the border width is constant, then the difference between these
two measurements is the difference between the scrollbar sizes at these two
states.

This can only detect a change in scrollbar size: it can't distinguish between
the case where a scrollbar exists during both measurements and the case where a
scrollbar does not exist during both measurements.

[^1]: This MDN page has some nice diagrams showing exactly what contributes to
each of `offsetWidth` and `clientWidth`:
https://developer.mozilla.org/en-US/docs/Web/API/CSS_Object_Model/Determining_the_dimensions_of_elements